move apart segmentation backbones
